Add files via upload
The Music Player project is an immersive and visually captivating web application that allows users to enjoy their favorite tunes in a seamless and interactive manner. Crafted with a perfect blend of HTML, CSS, and JavaScript, this music player aims to provide a delightful audio experience while showcasing an aesthetically pleasing user interface. This is a task for an internship of codeclause.